grc_articles/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-appconf==1.0.3
django-imagekit==4.0.2
djangorestframework==3.9.4
idna==2.8
pilkit==2.0
Pillow==6.0.0
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
